gui ve gmail

+++ app/models.py
@@ -87,7 +87,7 @@ class PaymentForm(forms.Form):
 
 class Payment_VNPay(models.Model):
     order_id = models.BigIntegerField(default=0, null=True, blank=True)
-    amount = models.FloatField(default=0.0, null=True, blank=True)
+    amount = models.BigIntegerField(default=0, null=True, blank=True)
     order_desc = models.CharField(max_length=200,null=True,blank=True)
     vnp_TransactionNo = models.CharField(max_length=200,null=True,blank=True)
     vnp_ResponseCode = models.CharField(max_length=200, null=True, blank=True)
